fix: adapt tests involving kt-math to new API from v0.6.2

+++ test-solve/src/commonTest/kotlin/it/unibo/tuprolog/solve/stdlib/function/FloatFractionalPartTest.kt
@@ -25,7 +25,7 @@ internal class FloatFractionalPartTest {
             val (_, fractionalPart) = parts
             assertEquals(
                 fractionalPart,
-                FloatFractionalPart.computeOf(Numeric.of(input)).castToReal().value.toDouble()
+                FloatFractionalPart.computeOf(Numeric.of(input)).castToReal().value
             )
         }
     }
